+++ src/Trarizon.Library/Collections/Helpers/SpanHelper/Creation/AsBytes.cs
@@ -6,18 +6,18 @@ namespace Trarizon.Library.Collections.Helpers;
 partial class SpanHelper
 {
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static unsafe Span<byte> AsBytes<T>(ref T value) where T : unmanaged
+    public static Span<byte> AsBytes<T>(ref T value) where T : unmanaged
     {
         return MemoryMarshal.CreateSpan(
             ref Unsafe.As<T, byte>(ref value),
-            sizeof(T));
+            Unsafe.SizeOf<T>());
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static unsafe ReadOnlySpan<byte> AsReadOnlyBytes<T>(ref readonly T value) where T : unmanaged
+    public static ReadOnlySpan<byte> AsReadOnlyBytes<T>(ref readonly T value) where T : unmanaged
     {
         return MemoryMarshal.CreateReadOnlySpan(
             in UnsafeHelper.AsReadOnly<T, byte>(in value),
-            sizeof(T));
+            Unsafe.SizeOf<T>());
     }
 }